+++ src/main/java/everTale/everTale_be/domain/alarm/dto/response/AlarmListResponseDto.java
@@ -2,9 +2,6 @@ package everTale.everTale_be.domain.alarm.dto.response;
 
 import everTale.everTale_be.domain.alarm.dto.AlarmSummary;
 import everTale.everTale_be.domain.alarm.entity.Alarm;
-import everTale.everTale_be.domain.story.dto.StoryCollectionResponseDto;
-import everTale.everTale_be.domain.story.dto.StorySummary;
-import everTale.everTale_be.domain.story.entity.Story;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
