+++ src/main/java/everTale/everTale_be/domain/story/dto/response/SceneResponseDTO.java
@@ -1,4 +1,4 @@
-package everTale.everTale_be.domain.story.dto;
+package everTale.everTale_be.domain.story.dto.response;
 
 import everTale.everTale_be.domain.story.entity.Scene;
 import lombok.Builder;
