+++ src/main/java/everTale/everTale_be/domain/story/service/StoryService.java
@@ -10,10 +10,10 @@ import everTale.everTale_be.domain.profile.entity.Enum.ProfileType;
 import everTale.everTale_be.domain.profile.entity.Profile;
 import everTale.everTale_be.domain.profile.service.ProfileService;
 import everTale.everTale_be.domain.profile.util.ProfileHelper;
-import everTale.everTale_be.domain.story.dto.SceneResponseDTO;
-import everTale.everTale_be.domain.story.dto.StoryCollectionResponseDto;
-import everTale.everTale_be.domain.story.dto.StoryRequestDTO;
-import everTale.everTale_be.domain.story.dto.StoryResponseDTO;
+import everTale.everTale_be.domain.story.dto.response.SceneResponseDTO;
+import everTale.everTale_be.domain.story.dto.response.StoryCollectionResponseDto;
+import everTale.everTale_be.domain.story.dto.request.StoryRequestDTO;
+import everTale.everTale_be.domain.story.dto.response.StoryResponseDTO;
 import everTale.everTale_be.domain.story.entity.Scene;
 import everTale.everTale_be.domain.story.entity.Story;
 import everTale.everTale_be.domain.story.repository.SceneRepository;
